Move config loading helpers into config folder.
