+++ packages/babel-core/src/config/helpers/environment.js
