+++ packages/babel-core/src/config/helpers/merge.js
