+++ packages/babel-core/src/config/helpers/resolve-plugin.js
