+++ packages/babel-core/test/get-possible-plugin-names.js
@@ -1,5 +1,5 @@
 import assert from "assert";
-import getPossiblePluginNames from "../lib/helpers/get-possible-plugin-names";
+import getPossiblePluginNames from "../lib/config/helpers/get-possible-plugin-names";
 
 describe("getPossiblePluginNames", function () {
   it("adds the babel-plugin prefix", function() {
